+++ Maccy/AppDelegate.swift
@@ -104,7 +104,7 @@ class AppDelegate: NSObject, NSApplicationDelegate {
 
     panel = FloatingPanel(
       contentRect: NSRect(origin: .zero, size: Defaults[.windowSize]),
-      identifier: Bundle.main.bundleIdentifier ?? "com.astrovini.maccy",
+      identifier: Bundle.main.bundleIdentifier ?? "com.astrovini.bettermaccy",
       statusBarButton: statusItem.button,
       onClose: { AppState.shared.popup.reset() }
     ) {